Перенёс viewClicks в класс Post, добавил в него свойство published
