+++ app/src/main/java/ru/netology/nmedia/MainActivity.kt
@@ -13,6 +13,7 @@ class MainActivity : AppCompatActivity() {
             1,
             getString(R.string.netologyName),
             getString(R.string.message),
+            getString(R.string.date),
             getString(R.string.likes),
             getString(R.string.shares)
 
@@ -28,29 +29,22 @@ class MainActivity : AppCompatActivity() {
                     binding.imageViewLikes.setImageResource(R.drawable.ic_baseline_thumb_up_alt_24)
                     post.isLiked = true
                     likes++
-                    binding.textViewLikes.setText(viewClicks(likes))
+                    binding.textViewLikes.setText(post.viewClicks(likes))
                 } else {
                     binding.imageViewLikes.setImageResource(R.drawable.ic_baseline_thumb_up_off_alt_24)
                     likes--
                     post.isLiked = false
-                    binding.textViewLikes.setText(viewClicks(likes))
+                    binding.textViewLikes.setText(post.viewClicks(likes))
                 }
             }
 
         }
         imageViewShare.setOnClickListener{
-            binding.textViewShares.setText(viewClicks(shares++))
+            binding.textViewShares.setText(post.viewClicks(shares++))
         }
 
     }
 
 
-    fun viewClicks(count: Int): String {
-        return when (count) {
-            in 0..1099 -> "$count"
-            in 1100..10_000 -> "${round((count / 100).toDouble()) / 10}K"
-            in 10_001..999_999 -> "${round((count / 1000).toDouble())}K"
-            else -> "${round((count / 100_000).toDouble()) / 10}M"
-        }
+
     }
-}
